pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simplemmo-bot"
version = "0.1.0"
description = "Automated travel and resource farming bot for SimpleMMO"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
simplemmo-bot = "simplemmo_bot.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
